Add SSM parameters for Github Actions
This commit adds SSM parameters to store secrets needed by Github Actions to build and push Docker images to ECR. The secrets include the Bedrock knowledge base name and ID, as well as the ECR registry and repository names. These parameters are created using Terraform and tagged for easy identification.